+++ tests/tests_pytorch/utilities/test_deepspeed_collate_checkpoint.py
@@ -1,4 +1,4 @@
-# Copyright The Lightning AI team.
+# Copyright The PyTorch Lightning team.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,10 +15,10 @@ import os
 
 import torch
 
-from lightning.pytorch import Trainer
-from lightning.pytorch.demos.boring_classes import BoringModel
-from lightning.pytorch.strategies import DeepSpeedStrategy
-from lightning.pytorch.utilities.deepspeed import convert_zero_checkpoint_to_fp32_state_dict
+from pytorch_lightning import Trainer
+from pytorch_lightning.demos.boring_classes import BoringModel
+from pytorch_lightning.strategies import DeepSpeedStrategy
+from pytorch_lightning.utilities.deepspeed import convert_zero_checkpoint_to_fp32_state_dict
 from tests_pytorch.helpers.runif import RunIf
 
 
@@ -32,7 +32,7 @@ def test_deepspeed_collate_checkpoint(tmpdir):
         accelerator="gpu",
         devices=2,
         fast_dev_run=True,
-        precision="16-mixed",
+        precision=16,
         enable_progress_bar=False,
         enable_model_summary=False,
     )
